chore: move enums to spi, use enums in JS API
