+++ src/commonMain/kotlin/org/hisp/dhis/lib/expression/spi/DataItemType.kt
@@ -1,11 +1,16 @@
 package org.hisp.dhis.lib.expression.spi
 
+import kotlin.js.ExperimentalJsExport
+import kotlin.js.JsExport
+
 /**
  * The different types of data value references in the expression grammar.
  *
  * @author Jan Bernitt
  */
-enum class DataItemType(val symbol: String, private val parameterTypes: List<List<IDType>>) {
+@OptIn(ExperimentalJsExport::class)
+@JsExport
+enum class DataItemType(internal val symbol: String, private val parameterTypes: List<List<IDType>>) {
 
     // (data element for aggregate vs. program stage . data element for programs)
     DATA_ELEMENT(
