+++ src/commonMain/kotlin/org/hisp/dhis/lib/expression/spi/Variable.kt
@@ -1,7 +1,5 @@
 package org.hisp.dhis.lib.expression.spi
 
-import org.hisp.dhis.lib.expression.ast.ProgramVariable
-
 data class Variable(
     val name: ProgramVariable,
     val modifiers: QueryModifiers
