+++ src/jsMain/kotlin/org.hisp.dhis.lib.expression.js/DataItemJs.kt
@@ -6,18 +6,9 @@ import org.hisp.dhis.lib.expression.spi.ID
 @OptIn(ExperimentalJsExport::class)
 @JsExport
 data class DataItemJs(
-    val type: String,
+    val type: DataItemType,
     val uid0: ID,
     val uid1: Array<ID>,
     val uid2: Array<ID>,
     val modifiers: QueryModifiersJs
-) {
-
-    init {
-        require(TYPES.contains(type)) { "DataItem type must be one of: $TYPES" }
-    }
-
-    companion object {
-        val TYPES = DataItemType.entries.map { it.name } .toTypedArray()
-    }
-}
+)
