+++ src/jsMain/kotlin/org.hisp.dhis.lib.expression.js/QueryModifiersJs.kt
@@ -1,13 +1,13 @@
 package org.hisp.dhis.lib.expression.js
 
 import kotlinx.datetime.LocalDate
-import org.hisp.dhis.lib.expression.ast.AggregationType
+import org.hisp.dhis.lib.expression.spi.AggregationType
 
 @OptIn(ExperimentalJsExport::class)
 @JsExport
 data class QueryModifiersJs(
     val periodAggregation: Boolean,
-    val aggregationType: String?,
+    val aggregationType: AggregationType?,
     val maxDate: String?,
     val minDate: String?,
     val periodOffset: Int?,
@@ -16,14 +16,11 @@ data class QueryModifiersJs(
     val subExpression: String? // SQL
 ) {
     init {
-        require(AGGREGATION_TYPES.contains(aggregationType)) { "QueryModifiers aggregationType must be one of: $AGGREGATION_TYPES" }
         requireDoesNotThrow(maxDate, LocalDate::parse)
         requireDoesNotThrow(minDate, LocalDate::parse)
     }
 
     companion object {
-        val AGGREGATION_TYPES = AggregationType.entries.map { it.name }.toTypedArray()
-
         internal fun <T> requireDoesNotThrow(value: T?, f: (T) -> Any?) {
             if (value == null) return
             try {
